requirements.txt
setuptools
python-crontab
PyYAML
requests
kubernetes
openshift
